requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
distro==1.9.0
Django==5.0.4
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.25.0
pydantic==2.7.1
pydantic_core==2.18.2
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
